requirements.txt
anthropic
pymongo
mcp[cli]
python-dotenv
requests


